Correct title visibility for story places
Two issues still seem to exist with story titles: A story place's <h1>
tag will sometimes get duplicated in place detail views, and it will
sometimes be invisible in the place list view. What we want is actually
the opposite behavior. Let's put an is-visuallyhidden class on story place
detail titles, and remove this class when the list view is shown.