Retain landmark titles when not part of a story
Because of the way we're parsing landmark HTML to pull out the title
for use in the story bar, landmarks which are not part of a story
still get their titles extracted and then appear title-less in their
detail views. Let's correct this by retaining a reference to the pre-parsed
landmark HTML and rendering that when a landmark is not part of a story.