Make sure attachment attribute is set in the list view for new place models
Creating a new place with an attachment will render the attachment in the
place detail view but not immediately in the place list view. The model used
by the place list view seems to lack the attachments attribute that is set
explicitly in the place detail view. Let's make sure it's set in the place list
view for new models as well.