lazy loading layers: Bind landmark collection listeners in initialize
Previously, landmark collection listeners were getting bound only
when the visibility event was triggered for a given landmark collection.
As a result, it was possible that landmark models would be fetched but
their layer views never created. To fix this, make sure we bind collection
listeners in the initialize method of the map view.

It was also convenient in this commit to support layer toggling for individual
shareabouts place collections. This does away with the separate shareabouts
layer in the layers section of the config.

Addresses: #436
Fixes: #408