On panel show, call invalidateSize() inside a short timeout
This is an effort to avoid a potential race condition between the
CSS transition that resizes the map container to accommodate the
content panel, and the call to invalidateSize() inside showPanel().
The timeout (hopefully) allows the transition to complete and the
new map dimensions to be set before invalidateSize() recenters the map.