Fix a performance bottleneck related to filtering
Because the viewLandmark route handler was not included on the list
of routes that don't require a clear of the location type filter, we
were running unnecessary filter clearing logic in routes.js and in
place-list-view.js whenever a landmark (or a landmark-as-place)
was clicked on.

This commit adds viewLandmark to the list of routes that don't
require filter clearing.